+++ backend/src/baserow/contrib/database/table/handler.py
@@ -142,10 +142,6 @@ class TableHandler:
             user_session_id=session,
             sync=sync,
         )
-
-        if sync:
-            job.refresh_from_db()
-
         return job
 
     def create_minimal_table(
@@ -209,9 +205,6 @@ class TableHandler:
             user_session_id=session,
             sync=sync,
         )
-
-        if sync:
-            job.refresh_from_db()
 
         return job